"Refactor OpenSSL initialization by replacing `sslinit` with `opensslgbl`, introducing a more modular and thread-safe design; adjust Makefiles and source files accordingly, and clean up unused includes."
