+++ src/Makefile
@@ -14,10 +14,12 @@ endif # EVENT2_HAS_OPENSSL
 USR_CPPFLAGS += -DPVXS_API_BUILDING
 USR_CPPFLAGS += -DPVXS_ENABLE_EXPERT_API
 
+ifeq ($(EVENT2_HAS_OPENSSL),YES)
 PVXS_ENABLE_SSLKEYLOGFILE ?= YES
 
 PVXS_ENABLE_SSLKEYLOGFILE_YES = -DPVXS_ENABLE_SSLKEYLOGFILE
 USR_CPPFLAGS += $(PVXS_ENABLE_SSLKEYLOGFILE_$(PVXS_ENABLE_SSLKEYLOGFILE))
+endif # EVENT2_HAS_OPENSSL
 
 ifdef T_A
 ifneq ($(CONFIG_LOADED),YES)
@@ -57,8 +59,10 @@ GENVERSIONMACRO = PVXS_VCS_VERSION
 SHRLIB_VERSION = $(PVXS_MAJOR_VERSION).$(PVXS_MINOR_VERSION)
 
 # Access to certs specific headers
+ifeq ($(EVENT2_HAS_OPENSSL),YES)
 USR_CPPFLAGS += -I$(TOP)/certs
 SRC_DIRS += $(TOP)/certs
+endif # EVENT2_HAS_OPENSSL
 
 INC += pvxs/client.h
 INC += pvxs/config.h
@@ -69,7 +73,6 @@ INC += pvxs/nt.h
 INC += pvxs/server.h
 INC += pvxs/sharedArray.h
 INC += pvxs/sharedpv.h
-INC += pvxs/sharedwildcardpv.h
 INC += pvxs/source.h
 INC += pvxs/srvcommon.h
 INC += pvxs/unittest.h
@@ -77,11 +80,13 @@ INC += pvxs/util.h
 INC += pvxs/version.h
 INC += pvxs/versionNum.h
 
+ifeq ($(EVENT2_HAS_OPENSSL),YES)
+INC += pvxs/sharedwildcardpv.h
+endif # EVENT2_HAS_OPENSSL
+
 LIBRARY = pvxs
 
 LIB_SRCS += bitmask.cpp
-LIB_SRCS += certstatus.cpp
-LIB_SRCS += certstatusmanager.cpp
 LIB_SRCS += client.cpp
 LIB_SRCS += clientconn.cpp
 LIB_SRCS += clientdiscover.cpp
@@ -98,8 +103,6 @@ LIB_SRCS += describe.cpp
 LIB_SRCS += evhelper.cpp
 LIB_SRCS += log.cpp
 LIB_SRCS += nt.cpp
-LIB_SRCS += openssl.cpp
-LIB_SRCS += sslinit.cpp
 LIB_SRCS += osdSockExt.cpp
 LIB_SRCS += osgroups.cpp
 LIB_SRCS += pvrequest.cpp
@@ -112,15 +115,22 @@ LIB_SRCS += servermon.cpp
 LIB_SRCS += serversource.cpp
 LIB_SRCS += sharedarray.cpp
 LIB_SRCS += sharedpv.cpp
-LIB_SRCS += sharedwildcardpv.cpp
 LIB_SRCS += type.cpp
 LIB_SRCS += udp_collector.cpp
 LIB_SRCS += unittest.cpp
 LIB_SRCS += util.cpp
 
+ifeq ($(EVENT2_HAS_OPENSSL),YES)
+LIB_SRCS += openssl.cpp
+LIB_SRCS += opensslgbl.cpp
+LIB_SRCS += certstatus.cpp
+LIB_SRCS += certstatusexdata.cpp
+LIB_SRCS += certstatusmanager.cpp
 LIB_SRCS += certfactory.cpp
 LIB_SRCS += certfilefactory.cpp
 LIB_SRCS += p12filefactory.cpp
+LIB_SRCS += sharedwildcardpv.cpp
+endif # EVENT2_HAS_OPENSSL
 
 LIB_LIBS += Com
