+++ src/certstatusexdata.cpp
@@ -1,0 +1,9 @@
+#include "certstatusexdata.h"
+
+namespace pvxs {
+namespace ossl {
+
+CertStatusExDataBase::~CertStatusExDataBase() noexcept = default;
+
+} // namespace ossl
+} // namespace pvxs
