+++ src/certstatusexdata.h
@@ -1,0 +1,14 @@
+#ifndef PVXS_CERTESTATUSEXDATA_H
+#define PVXS_CERTESTATUSEXDATA_H
+
+namespace pvxs {
+namespace ossl {
+
+struct CertStatusExDataBase {
+    virtual ~CertStatusExDataBase() noexcept;
+};
+
+} // namespace ossl
+} // namespace pvxs
+
+#endif
