+++ tools/call.cpp
@@ -7,20 +7,13 @@
 #include <iostream>
 #include <map>
 #include <list>
-#include <atomic>
 
-#include <epicsVersion.h>
-#include <epicsGetopt.h>
 #include <epicsThread.h>
 
 #include <pvxs/client.h>
 #include <pvxs/nt.h>
 #include <pvxs/log.h>
-#ifdef PVXS_ENABLE_OPENSSL
-#include <pvxs/sslinit.h>
-#endif
 #include "utilpvt.h"
-#include "evhelper.h"
 
 using namespace pvxs;
 
@@ -43,9 +36,6 @@ void usage(const char* argv0)
 int main(int argc, char *argv[])
 {
     try {
-#ifdef PVXS_ENABLE_OPENSSL
-        ossl::sslInit();
-#endif
         logger_config_env(); // from $PVXS_LOG
         double timeout = 5.0;
         bool verbose = false;
