+++ tools/list.cpp
@@ -18,9 +18,6 @@
 #include <pvxs/client.h>
 #include <pvxs/nt.h>
 #include <pvxs/log.h>
-#ifdef PVXS_ENABLE_OPENSSL
-#include <pvxs/sslinit.h>
-#endif
 
 #include "utilpvt.h"
 #include "evhelper.h"
@@ -68,9 +65,6 @@ void usage(const char* argv0)
 int main(int argc, char *argv[])
 {
     try {
-#ifdef PVXS_ENABLE_OPENSSL
-        ossl::sslInit();
-#endif
         logger_config_env(); // from $PVXS_LOG
         double timeout = 5.0;
         bool verbose = false;
