+++ tools/pvxvct.cpp
@@ -22,9 +22,6 @@
 #include <osiSock.h>
 
 #include <pvxs/log.h>
-#ifdef PVXS_ENABLE_OPENSSL
-#include <pvxs/sslinit.h>
-#endif
 #include <pvxs/util.h>
 
 #include <udp_collector.h>
@@ -105,9 +102,6 @@ void usage(const char *name)
 int main(int argc, char *argv[])
 {
     try {
-#ifdef PVXS_ENABLE_OPENSSL
-        pvxs::ossl::sslInit();
-#endif
         // group options used from callback
         struct {
             bool verbose = false;
